Move ipv8_health_manager
