+++ src/tribler-core/tribler_core/components/ipv8/ipv8_component.py
@@ -69,7 +69,7 @@ class Ipv8Component(RestfulComponent):
                 ipv8.endpoint.enable_community_statistics(overlay.get_prefix(), True)
 
         if config.ipv8.walk_scaling_enabled and not config.gui_test_mode:
-            from tribler_core.modules.ipv8_health_monitor import IPv8Monitor
+            from tribler_core.components.ipv8.ipv8_health_monitor import IPv8Monitor
             IPv8Monitor(ipv8,
                         config.ipv8.walk_interval,
                         config.ipv8.walk_scaling_upper_limit).start(self._task_manager)
